requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.5
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
future==0.18.2
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
